Small adjustments to speed up creating the collection Minor bugfixes Major change - collection is now cached so will be loaded much faster. New option to make it update every time you load the page and new buttons to force an update or a rescan
git-svn-id: http://svn.code.sf.net/p/rompr/code/trunk@124 94fc339e-1912-44e7-90be-6e134bb66470

+++ connection.php
@@ -76,7 +76,8 @@ function clean_stored_xspf() {
 
     $playlists = glob("prefs/*.xspf");
     foreach($playlists as $i => $file) {
-        if (!preg_match('/USERSTREAM/', basename($file))) {
+        if (!preg_match('/USERSTREAM/', basename($file)) &&
+            ($file != "prefs/icecast.xspf")) {
             system("rm ".$file);
         }
     }
